+++ webapp-v4/apps/reception/components/ReceptionRoot.vue
@@ -6,6 +6,33 @@
 
 <template>
   <div>
+    <!-- The dialog for warnings using a too low screen resolution -->
+    <Dialog v-model:visible="dialogVisible" modal >
+      <template #container="{ closeCallback }">
+        <div class="flex flex-col items-center p-8 bg-surface-0 dark:bg-surface-900 rounded">
+          <div
+              class="rounded-full bg-primary text-primary-contrast inline-flex justify-center items-center h-24 w-24 -mt-20"
+          >
+            <FontAwesomeIcon
+                id="confirm-icon"
+                :icon="faDisplay"
+            />
+          </div>
+          <span class="font-bold text-2xl block mb-2 mt-6">Houston, we have a problem...</span>
+          <p class="mb-0">Die Auflösung deines Bildschirms ist zu klein - die Spielauswertung ist ausgelegt für Bildschirme mit
+            &#8805; 1024 Pixel, mit dem Handy lässt sich das Spiel sehr schlecht administrieren.</p>
+          <div class="flex items-center gap-2 mt-6">
+            <Button
+                label="Verstanden, ich mach trotzdem weiter"
+                class="w-auto"
+                @click="closeCallback"
+            />
+          </div>
+
+        </div>
+      </template>
+    </Dialog>
+    <!-- here starts the app -->
     <menu-bar
         :elements="receptionStore.menuBarElements"
         show-online-status
@@ -26,6 +53,11 @@ import {last, split,get} from 'lodash';
 import {getReceptionSocket} from '../lib/ReceptionSocket';
 import {useTeamsStore} from '../../../lib/store/TeamsStore';
 import {usePropertyStore} from '../../../lib/store/PropertyStore';
+import {onMounted, ref} from 'vue';
+import Dialog from 'primevue/dialog';
+import {faDisplay} from '@fortawesome/free-solid-svg-icons';
+import {FontAwesomeIcon} from '@fortawesome/vue-fontawesome';
+import Button from 'primevue/button';
 
 const receptionStore = useReceptionStore();
 const receptionSocket = getReceptionSocket();
@@ -51,8 +83,22 @@ receptionStore.fetchStaticData(gameId)
       console.error(err);
     });
 
+const dialogVisible = ref(false);
+
+onMounted(()=> {
+  console.log('window.screen.width', window.screen.width);
+  if (window.screen.width < 1024) {
+    dialogVisible.value = true;
+  }
+})
+
+const closeCallback = function() {
+  dialogVisible.value = false;
+}
 </script>
 
 <style scoped lang="scss">
-
+#confirm-icon {
+  font-size: xxx-large;
+}
 </style>